Adjust appearance for laika testnet purpose.
Due to the upstream merge, the layout and colorscheme has been changed.
There the old logo has been exchanged and add a new footer logo.
Moreover the colors have been adapted to fit for the new scheme. Also
the colors are now related to the trustlines protocol instead of
*-network.

Related to: brainbot-com/trustlines-meta#114
See also: #8